Added Shooting, and the start of basic Finite State Machines
Fixed and tested the shooting component to function as intended. This also includes making a TryShoot function, which puts a limit on how many shots can be fired within a second. Also began adding some of the Finite State Machine AI mechanics, however this is still unfinished until the FSM diagrams of my AI behaviors are finished.